facturacion_ocr_app/requirements.txt
Flask
flask-cors
python-dotenv
requests
pillow
tesserocr